RavenDB-10221 - disable setting the working set
